WORK/Outils Production/SEO AUTOMATION TOOLKIT - V2/requirements.txt
requests
python-dotenv
google-api-python-client
lxml
